# Replace Stack with ArrayBuilder in GreenNode walker