--- a/src/Compilers/Core/Portable/Syntax/GreenNode.cs
+++ b/src/Compilers/Core/Portable/Syntax/GreenNode.cs
@@ -644,47 +644,50 @@ namespace Microsoft.CodeAnalysis
         protected internal void WriteTo(TextWriter writer, bool leading, bool trailing)
         {
             // Use an actual Stack so we can write out deeply recursive structures without overflowing.
-            var stack = new Stack<(GreenNode node, bool leading, bool trailing)>();
+            var stack = ArrayBuilder<(GreenNode node, bool leading, bool trailing)>.GetInstance();
             stack.Push((this, leading, trailing));
 
             // Separated out stack processing logic so that it does not unintentionally refer to 
             // "this", "leading" or "trailing.
-            ProcessStack(writer, stack);
-        }
+            processStack(writer, stack);
+            stack.Free();
+            return;
 
-        private static void ProcessStack(TextWriter writer,
-            Stack<(GreenNode node, bool leading, bool trailing)> stack)
-        {
-            while (stack.Count > 0)
+            static void processStack(
+                TextWriter writer,
+                ArrayBuilder<(GreenNode node, bool leading, bool trailing)> stack)
             {
-                var current = stack.Pop();
-                var currentNode = current.node;
-                var currentLeading = current.leading;
-                var currentTrailing = current.trailing;
-
-                if (currentNode.IsToken)
+                while (stack.Count > 0)
                 {
-                    currentNode.WriteTokenTo(writer, currentLeading, currentTrailing);
-                    continue;
-                }
+                    var current = stack.Pop();
+                    var currentNode = current.node;
+                    var currentLeading = current.leading;
+                    var currentTrailing = current.trailing;
 
-                if (currentNode.IsTrivia)
-                {
-                    currentNode.WriteTriviaTo(writer);
-                    continue;
-                }
-
-                var firstIndex = GetFirstNonNullChildIndex(currentNode);
-                var lastIndex = GetLastNonNullChildIndex(currentNode);
-
-                for (var i = lastIndex; i >= firstIndex; i--)
-                {
-                    var child = currentNode.GetSlot(i);
-                    if (child != null)
+                    if (currentNode.IsToken)
                     {
-                        var first = i == firstIndex;
-                        var last = i == lastIndex;
-                        stack.Push((child, currentLeading | !first, currentTrailing | !last));
+                        currentNode.WriteTokenTo(writer, currentLeading, currentTrailing);
+                        continue;
+                    }
+
+                    if (currentNode.IsTrivia)
+                    {
+                        currentNode.WriteTriviaTo(writer);
+                        continue;
+                    }
+
+                    var firstIndex = GetFirstNonNullChildIndex(currentNode);
+                    var lastIndex = GetLastNonNullChildIndex(currentNode);
+
+                    for (var i = lastIndex; i >= firstIndex; i--)
+                    {
+                        var child = currentNode.GetSlot(i);
+                        if (child != null)
+                        {
+                            var first = i == firstIndex;
+                            var last = i == lastIndex;
+                            stack.Push((child, currentLeading | !first, currentTrailing | !last));
+                        }
                     }
                 }
             }