--- a/src/Compilers/Core/Portable/Syntax/GreenNode.cs
+++ b/src/Compilers/Core/Portable/Syntax/GreenNode.cs
@@ -643,12 +643,12 @@ namespace Microsoft.CodeAnalysis
 
         protected internal void WriteTo(TextWriter writer, bool leading, bool trailing)
         {
-            // Use an actual Stack so we can write out deeply recursive structures without overflowing.
+            // Use an actual stack so we can write out deeply recursive structures without overflowing.
             var stack = ArrayBuilder<(GreenNode node, bool leading, bool trailing)>.GetInstance();
             stack.Push((this, leading, trailing));
 
             // Separated out stack processing logic so that it does not unintentionally refer to 
-            // "this", "leading" or "trailing.
+            // "this", "leading" or "trailing".
             processStack(writer, stack);
             stack.Free();
             return;