@jcouv 2019-03-04
Thanks for remembering about this issue :-) 
Could you also share the command you use to run those benchmarks?